--- src/main/java/com/solvd/carina/demo/gui/pages/demoqa/constant/ElementsConstant.java
@@ -1,9 +1,0 @@
-package com.solvd.carina.demo.gui.pages.demoqa.constant;
-
-import org.openqa.selenium.By;
-
-public class ElementsConstant {
-
-    public static final By RESULT_INFO_FIELDS = By.xpath("//div[@class='border col-md-12 col-sm-12']/p[@id]");
-    public static final By FOLDERS = By.xpath("//div[@id='tree-node']/ol//li[@class='rct-node rct-node-leaf']");
-}
