--- src/main/java/com/solvd/carina/demo/gui/pages/shop24/constant/PageLocatorsConstants.java
@@ -1,7 +1,0 @@
-package com.solvd.carina.demo.gui.pages.shop24.constant;
-
-import org.openqa.selenium.By;
-
-public class PageLocatorsConstants {
-    public static final By MESSAGE_SUCCESS_ADDING_BASKET = By.className("pp-added");
-}
