+++ src/main/java/com/solvd/shop24/gui/pages/components/shop24/constant/LocatorsConstants.java
@@ -1,0 +1,11 @@
+package com.solvd.shop24.gui.pages.components.shop24.constant;
+
+import org.openqa.selenium.By;
+
+public class LocatorsConstants {
+
+    public static final By BASKET_BUTTON = By.xpath("//div[@class='go-to-basket']/a");
+    public static final By OPENED_CATALOG_MENU = By.className("catalog-btn open");
+
+    private LocatorsConstants() {}
+}
