+++ src/main/java/com/solvd/shop24/gui/pages/shop24/constant/PageLocatorsConstants.java
@@ -1,0 +1,7 @@
+package com.solvd.shop24.gui.pages.shop24.constant;
+
+import org.openqa.selenium.By;
+
+public class PageLocatorsConstants {
+    public static final By MESSAGE_SUCCESS_ADDING_BASKET = By.className("pp-added");
+}
