+++ src/main/java/com/solvd/shop24/gui/pages/shop24/news/NewsPage.java
@@ -1,0 +1,34 @@
+package com.solvd.shop24.gui.pages.shop24.news;
+
+import com.qaprosoft.carina.core.foundation.webdriver.decorator.ExtendedWebElement;
+import com.solvd.shop24.gui.pages.components.shop24.ArticleItem;
+import com.solvd.shop24.gui.pages.shop24.HomePage;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.support.FindBy;
+
+import java.util.List;
+
+public class NewsPage extends HomePage {
+
+    @FindBy(className = "page-main__title")
+    private ExtendedWebElement title;
+
+    @FindBy(className = "news-item__content")
+    private List<ArticleItem> articlesList;
+
+    public NewsPage(WebDriver driver) {
+        super(driver);
+    }
+
+    public String getTitle() {
+        return title.getText();
+    }
+
+    public List<ArticleItem> getArticlesList() {
+        return articlesList;
+    }
+
+    public ArticleItem getArticle(int index) {
+        return articlesList.get(index);
+    }
+}
